Add check script
My intent here is that you'd run this before pushing a commit to make
sure you didn't forget anything. I'm new to this project so I'm not sure
if I'm missing any checks that should be in here.

Currently it checks that running the regenerate scripts doesn't change
anything and that all the tests pass.